Factor out server_client into cheroot.testing
