+++ cheroot/test/conftest.py
@@ -1,2 +1,2 @@
 """Pytest configuration module."""
-from ..testing import native_server, wsgi_server  # noqa: F401
+from ..testing import native_server, wsgi_server, server_client  # noqa: F401
